move src/write_all.rs -> src/io/write_all.rs
